# constant time field.Normalize()