--- a/btcec/field.go
+++ b/btcec/field.go
@@ -55,6 +55,7 @@ package btcec
 // counts.
 
 import (
+	"crypto/subtle"
 	"encoding/hex"
 )
 
@@ -319,32 +320,12 @@ func (f *fieldVal) Normalize() *fieldVal {
 	// following determines if either or these conditions are true and does
 	// the final reduction in constant time.
 	//
-	// Note that the if/else statements here intentionally do the bitwise
-	// operators even when it won't change the value to ensure constant time
-	// between the branches.  Also note that 'm' will be zero when neither
-	// of the aforementioned conditions are true and the value will not be
-	// changed when 'm' is zero.
-	m = 1
-	if t9 == fieldMSBMask {
-		m &= 1
-	} else {
-		m &= 0
-	}
-	if t2&t3&t4&t5&t6&t7&t8 == fieldBaseMask {
-		m &= 1
-	} else {
-		m &= 0
-	}
-	if ((t0+977)>>fieldBase + t1 + 64) > fieldBaseMask {
-		m &= 1
-	} else {
-		m &= 0
-	}
-	if t9>>fieldMSBBits != 0 {
-		m |= 1
-	} else {
-		m |= 0
-	}
+	// Note that 'm' will be zero when neither of the aforementioned conditions
+	// are true and the value will not be changed when 'm' is zero.
+	m = uint32(subtle.ConstantTimeEq(int32(t9), fieldMSBMask))
+	m &= uint32(subtle.ConstantTimeEq(int32(t2&t3&t4&t5&t6&t7&t8), fieldBaseMask))
+	m &= uint32(subtle.ConstantTimeLessOrEq(1<<fieldBase, int((t0+977)>>fieldBase+t1+64)))
+	m |= t9 >> fieldMSBBits
 	t0 = t0 + m*977
 	t1 = (t0 >> fieldBase) + t1 + (m << 6)
 	t0 = t0 & fieldBaseMask