The current implementation of `field.Normalize()` has a timing vulnerability exposed in issue #1079. This PR solves that issue by using methods outlined in the crypto/subtle reference library. 

Several constant time comparison operators that return integers for use in further bitwise operations are defined and should be used in place of other branching statements in the btcec library (even if-else statements with equal-time branches are vulnerable to timing attacks based on branch-prediction!). I think this solves #1079, but please comment if you knows of any other place following this same-time branching statements pattern. For details comparing the performance of these functions to their same-time branching counterparts see [my example repo](https://github.com/bmperrea/constant-time-go). For a more full set of constant time functions see [this other branch](https://github.com/bmperrea/testing_nfield).

Credits to @stevenroose for questioning this pattern in #621. Also, a plug for everyone to suggest there be a direct `bool2int` conversion function in golang!