@bmperrea 2017-11-27
```
benchmark                          old ns/op     new ns/op     delta
-----------------------------------------------------------------------
BenchmarkAddJacobian-8             561           597           +6.42%
BenchmarkAddJacobianNotZOne-8      1158          1176          +1.55%
BenchmarkScalarBaseMult-8          50553         51160         +1.20%
BenchmarkScalarBaseMultLarge-8     50779         52366         +3.13%
BenchmarkScalarMult-8              166077        172265        +3.73%
BenchmarkNAF-8                     743           742           -0.13%
BenchmarkSigVerify-8               265149        270769        +2.12%
BenchmarkFieldNormalize-8          14.2          20.4          +43.66%
BenchmarkFieldNormalize2-8         19.4          20.4          +5.15%
```
Unfortunately, the slowdown here is larger than the speedup found in [the most recent change to this file](https://github.com/btcsuite/btcd/commit/1238b7e55abe995643520aeb6b52b47b98d5d13f). However, as demonstrated in `BenchmarkFieldNormalize2` basically all of this slowdown is due to the faster runtime of the average case of `Normalize` when using branching, and therefore the slowdown unavoidable if the runtime is to be truly constant.