@stevenroose 2017-11-27
Thanks for the credits :)

Cool that you managed to do a proper constant time, I struggled with it at that time!

This change would be compatible with https://github.com/btcsuite/btcd/pull/621, right?