@stevenroose 2017-11-27
Also, you seem to have forgotten to run gofmt. It's always a good idea to run the `goclean.sh` command before committing, the Travis here runs it as well.