@bmperrea 2017-11-27
Thanks @stevenroose. It looks like that fixed the tests.

This is compatible with my suggested branch for #621, although it will probably require a simple rebase. If I am able to get this reviewed and merged then I would want to spend some more time with #621 to improve its performance and get it finished off as well. Great work on this stuff btw.