@stevenroose 2017-11-28
Wait until Schnorr arrives, we're gonna need you here! :)