@bmperrea 2017-11-28
I added a commit with where I change the internals of `constant_time` to use `int32` instead of `int64`. I realized we know that the 10 `uint32` values in `field` will be less than 2^27 and thought that `int32` might be more consistently performant on different architectures.