@bmperrea 2017-12-02
Thanks for the excellent review @davecgh. I have made the updates, squashed, and renamed the commit.