@bmperrea 2019-11-18
Updated based on good feedback in https://github.com/gcash/bchd/pull/78 . The timing got faster since doing so - thanks @jimpo and @markblundeberg